renders new chats without bugs

+++ client/src/components/Recipients.jsx
@@ -3,21 +3,23 @@ import Chip from '@material-ui/core/Chip';
 import Paper from '@material-ui/core/Paper';
 
 
-const Recipients = ({ recipients, setRecipients }) => {
-
+const Recipients = ({ recipients, setRecipients, user: currentUser }) => {
   const deleteRecipient = (recipient) => {
-    const recipientCopy = recipients.filter(user => user._id !== recipient._id);
+    const recipientCopy = recipients.filter((user) => user._id !== recipient._id);
     setRecipients(recipientCopy);
   };
-
-  const chips = recipients.map((user) => (
-    <Chip
-      style={{ padding: 3 }}
-      key={user._id}
-      label={user.name}
-      onDelete={() => deleteRecipient(user)}
-    />
-  ));
+  console.log(recipients, 'quuiiii');
+  const chips = recipients.map((user) => {
+    if (!user._id || user._id === currentUser._id) return null;
+    return (
+      <Chip
+        style={{ padding: 3 }}
+        key={user._id}
+        label={user.name}
+        onDelete={() => deleteRecipient(user)}
+      />
+    );
+  });
   return (
     <Paper>
       {chips}
